% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{analyze_HLA_data}
\alias{analyze_HLA_data}
\title{Analyze HLA typing files}
\usage{
analyze_HLA_data(HLA_data)
}
\arguments{
\item{HLA_data}{A container object of class HLA_data.}
}
\value{
A container object of class HLA_data, updated with analysis results.
}
\description{
Use negative and positive control files to determine
thresholds for all HLA channels. Then classify cells from all samples.
}
