% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{make_output_plot}
\alias{make_output_plot}
\title{Plot calls overlaid on data distribution for all files}
\usage{
make_output_plot(HLA_data)
}
\arguments{
\item{HLA_data}{A container object of class HLA_data.}
}
\value{
A ggplot2 object.
}
\description{
Shows scaled distributions of all channels in all files,
with calls and number of events overlaid.
}
