% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{plot_bivariate}
\alias{plot_bivariate}
\title{Bivariate plots of all cells}
\usage{
plot_bivariate(HLA_data, dir_out)
}
\arguments{
\item{HLA_data}{A container object of class HLA_data.}

\item{dir_out}{Path for writing plots.}
}
\value{
Nothing.
}
\description{
Produces bivariate plots color coded by HLA type of cells,
then saves them to file.
}
