% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap of subject HLA types}
\usage{
plot_heatmap(HLA_data)
}
\arguments{
\item{HLA_data}{A container object of class HLA_data.}
}
\value{
A heatmap ggplot2 object.
}
\description{
Shows percentage of cells in each sample having each HLA type.
}
