% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{plot_thresholds}
\alias{plot_thresholds}
\title{Plot channel distributions and thresholds}
\usage{
plot_thresholds(HLA_data)
}
\arguments{
\item{HLA_data}{A container object of class HLA_data.}
}
\value{
A ggplot2 object.
}
\description{
Shows position of thresholds relative to distribution
of negative controls, positive controls and aggregte data from all files.
}
