% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{plot_umap}
\alias{plot_umap}
\title{UMAP plot of all cells}
\usage{
plot_umap(HLA_data, dir_out)
}
\arguments{
\item{HLA_data}{A container object of class HLA_data.}

\item{dir_out}{Path for writing plots.}
}
\value{
Nothing.
}
\description{
Produces UMAP plots color coded by expression of all markers
and by HLA type of cells, then saves them to file.
}
