% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA.R
\name{read_HLA_data}
\alias{read_HLA_data}
\title{Parse HLA typing files}
\usage{
read_HLA_data(
  dir_in,
  channels = c("A2", "A3", "B7"),
  cofactor = 500,
  scale_scat = 40000
)
}
\arguments{
\item{dir_in}{A path to read fcs files from.}

\item{channels}{Fluorescence channels to be used in analysis.}

\item{cofactor}{Parameter for arcsinh transformation of fluorescence channels.}

\item{scale_scat}{Parameter for linear scaling of scattering channels.}
}
\value{
A container object of class HLA_data.
}
\description{
Read all fcs files in provided directory, apply
compensation and arcsinh transformation, and prepare for analysis.
}
